Práctica finita

+++ main.c
@@ -1,53 +1,64 @@
 #include "inc.h"
-/*
-    Instrucción main crea procesos y comprueba señales.
-*/
 
+// Estructura de datos que guarda los pids de la votación y los votos
 typedef struct _process_info process_info;
-
+// Manejador de la señal SIGINT
 void int_sig_handler(int sig)
 {
-    if (sem_trywait(sem_print) == 0)
+    if (getpid() == MAIN_PID)
     {
         FLAG = 2;
         return;
     }
     else
     {
+        
         exit(EXIT_FAILURE);
     }
 }
+// Manejador de la señal SIGUSR2
+
 void usr2_sig_handler(int sig)
 {
     return;
 }
-void alarm_sig_handler(int sig)
+// Manejador de la señal SIGUSR1
+
+void usr1_handler(int sig)
 {
-    if (sem_trywait(sem_print) == 0)
-    {
-        FLAG = 3;
-    }
     return;
 }
+// Manejador de la señal SIGALRM
+void alarm_sig_handler(int sig)
+{
+
+    FLAG = 3;
+    return;
+}
+// Mascaras necesarias para el funcionamiento del programa
 void mascaras()
 {
+    sigemptyset(&oset);
+    sigemptyset(&nomask);
     sigemptyset(&set);
     sigaddset(&set, SIGUSR1);
     sigaddset(&set, SIGUSR2);
+    sigaddset(&set, SIGALRM);
+    sigaddset(&set, SIGINT);
+
     if (sigprocmask(SIG_BLOCK, &set, &oset) < 0)
     {
         perror("sigprocmask");
         exit(EXIT_FAILURE);
     }
 }
-
+// Manejadores necesarios
 void manejadores()
 {
     // SIGUSR 1 manejador
     usr1_han.sa_handler = usr1_handler;
     sigemptyset(&(usr1_han.sa_mask));
     sigaddset(&(usr1_han.sa_mask), SIGUSR2);
-    sigaddset(&(usr1_han.sa_mask), SIGINT);
     usr1_han.sa_flags = 0;
     if (sigaction(SIGUSR1, &usr1_han, NULL) < 0)
     {
@@ -70,6 +81,7 @@ void manejadores()
 
     usr2_han.sa_handler = usr2_sig_handler;
     sigemptyset(&(usr2_han.sa_mask));
+    sigaddset(&(usr1_han.sa_mask), SIGUSR1);
     usr2_han.sa_flags = 0;
 
     if (sigaction(SIGUSR2, &usr2_han, NULL) < 0)
@@ -78,7 +90,7 @@ void manejadores()
         exit(EXIT_FAILURE);
     }
 
-    // ALARM HANDLER
+    // ALARM MANEJADOR
     alarm_han.sa_handler = alarm_sig_handler;
     sigemptyset(&(alarm_han.sa_mask));
     alarm_han.sa_flags = 0;
@@ -88,7 +100,7 @@ void manejadores()
         exit(EXIT_FAILURE);
     }
 }
-
+// SEMAFOROS UTILIZADOS EN EL PROGRAMA
 void semaforos()
 {
     if ((sem = sem_open("Create_Candidato", O_CREAT | O_EXCL, S_IRUSR | S_IWUSR, 1)) == SEM_FAILED)
@@ -101,43 +113,46 @@ void semaforos()
         perror("sem_create");
         exit(EXIT_FAILURE);
     }
-    if ((sem_print = sem_open("Print_Final", O_CREAT | O_EXCL, S_IRUSR | S_IWUSR, 1)) == SEM_FAILED)
-    {
-        perror("sem_create");
-        exit(EXIT_FAILURE);
-    }
 }
 
 int main(int argc, char *argv[])
 {
-
+    // creación de variables útiles
     pid_t proceso_act;
     FILE *procesos = NULL;
     process_info pinf;
+    int bytes = 0;
+
     /*
-        Comprobación de argumentos entregados
+        Comprobación de argumentos entregados y errores
     */
     if (argc != 3)
     {
         printf("El número de argumentos es incorrecto.");
         return EXIT_FAILURE;
     }
+    // Asignación de variables globales
     N_PROCESOS = atoi(argv[1]);
     N_SECS = atof(argv[2]);
     FLAG = 1;
     CONTADOR = 0;
-    printf("Main: %i\n",getpid());
+    MAIN_PID = getpid();
     pinf.pid_Arr = (pid_t *)malloc(sizeof(pid_t) * N_PROCESOS);
-
+    if (!pinf.pid_Arr)
+        return EXIT_FAILURE;
+    // Comprobación argumentos
     if (N_PROCESOS < 2 || N_SECS <= 0)
     {
         printf("Valores no validos de argumentos.");
         return EXIT_FAILURE;
     }
+    // Comprobación alarma
     if (alarm(N_SECS))
     {
         printf("Esa alarma ya existe");
+        return EXIT_FAILURE;
     }
+    // Creación y comprobación de fichero
     procesos = fopen("procesos.bin", "wb");
 
     if (!procesos)
@@ -146,14 +161,14 @@ int main(int argc, char *argv[])
         return EXIT_FAILURE;
     }
 
-    // HASTA AQUI BIEN
+    // Se crean las mascaras, manejadores y semaforos
     manejadores();
     mascaras();
     semaforos();
 
-    // HASTA AQUÍ BIEN
     for (int i = 0; i < N_PROCESOS; i++)
     {
+        // Se crean todos los procesos y llaman a la función main_votante()
         if ((proceso_act = fork()) == 0)
         {
             fflush(stdout);
@@ -161,46 +176,59 @@ int main(int argc, char *argv[])
         }
         else
         {
+            // Proceso padre añade a la estructura los pids de los procesos
             pinf.pid_Arr[i] = proceso_act;
         }
     }
-    // HASTA AQUI BIEN
-    fwrite(pinf.pid_Arr, sizeof(int), N_PROCESOS, procesos);
-    fflush(procesos);
-    fclose(procesos);
-    // HASTA AQUI BIEN
-    while (FLAG == 1)
+    // Escribe la estructura en un fichero binario
+    if ((bytes = fwrite(pinf.pid_Arr, sizeof(int), N_PROCESOS, procesos)) != N_PROCESOS)
     {
-        for (int i = 0; i < N_PROCESOS; i++)
-        {
-            fflush(stdout);
-            kill(pinf.pid_Arr[i], SIGUSR1);
-        }
-        sigsuspend(&oset);
+        printf("%i,\n", bytes);
+        free(pinf.pid_Arr);
+        sem_close(sem);
+        sem_close(sem_vot);
+        sem_unlink("Create_Candidato");
+        sem_unlink("Votar_Candidato");
+        return EXIT_FAILURE;
     }
+    fflush(procesos);
+    // Cierra el fichero
+    fclose(procesos);
+
+    // Se envia señal
     for (int i = 0; i < N_PROCESOS; i++)
     {
-        kill(pinf.pid_Arr[i], SIGKILL);
+        fflush(stdout);
+        // se envia una señal a los procesos creados
+        kill(pinf.pid_Arr[i], SIGUSR1);
     }
+    // Se suspende el proceso main hasta recibir señal
+    sigsuspend(&oset);
+    // Se ha mandado la señal SIGINT o SIGALARM asi que se sale y se envia la señal al resto de procesos
+    for (int i = 0; i < N_PROCESOS; i++)
+    {
 
+        kill(pinf.pid_Arr[i], SIGINT);
+    }
+    // Se espera a que finalicen los procesos
     for (int i = 0; i < N_PROCESOS; i++)
     {
         waitpid(pinf.pid_Arr[i], NULL, 0);
     }
-
+    // FLAG que indica que ha salido por señal
     if (FLAG == 2)
     {
         printf("Terminated by signal\n");
     }
+    // FLAG que indica que ha salido por alarma
     else
     {
         printf("Terminated by alarm\n");
     }
+    // Liberación de recursos.
     free(pinf.pid_Arr);
     sem_close(sem);
     sem_close(sem_vot);
-    sem_close(sem_print);
-    sem_unlink("Print_Final");
     sem_unlink("Create_Candidato");
     sem_unlink("Votar_Candidato");
